requirements.txt
streamlit
pdf2image
google.generativeai
python.dotenv
base64